pyproject.toml
[tool.poetry]
name = "todo-cli"
version = "0.1.0"
description = ""
authors = ["Ajith Aravind <ajith.aravind100@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"
SQLAlchemy = "^1.4.40"

[tool.poetry.dev-dependencies]
pylint = "^2.15.0"
autopep8 = "^1.7.0"
pytest = "^7.1.3"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.autopep8]
max_line_length = 120
in-place = true
recursive = true
aggressive = 3
